Test class for OutputDataItem
This class is for testing the updateResponseForLiteralData() method of
the class OutputDataItem.java. You will see that while creating each of
the currently supported LiteralDataBinding-outputs the message: "[Fatal
Error] :1:1: Content is not allowed in prolog." is logged. This happens
because it is tried to parse the literal data values as a XMLObject. Do
you see situations, in which the parsing actually makes sense? XML in
literal string outputs perhaps? Otherwise, I would have a fix
ready,which would be my next pull request.